+++ inst/template/app.R
@@ -35,18 +35,14 @@ ui <- sd_ui()
 # Server setup ----------------------------------------------------------------
 
 server <- function(input, output, session) {
-
   # Define any conditional skip logic here (skip to page if a condition is true)
   sd_skip_forward()
 
   # Define any conditional display logic here (show a question if a condition is true)
   sd_show_if()
 
-  # Database designation and other settings
-  sd_server(
-    db = db
-  )
-
+  # Run surveydown server and define database
+  sd_server(db = db)
 }
 
 # Launch the app